renaming code directory to lib to avoid osx compile issue
