+++ lib/bam_io.cpp
