+++ lib/bam_io.h
