+++ lib/bingenerator.cpp
