+++ lib/bingenerator.h
