+++ lib/common.cpp
