+++ lib/library_constructor.cpp
