+++ lib/library_constructor.h
