+++ lib/options.cpp
