+++ lib/peak_intervals.cpp
