+++ lib/peak_intervals.h
