+++ lib/sequencer.cpp
