+++ lib/sequencer.h
