+++ lib/stringops.cpp
