+++ lib/stringops.h
