+++ lib/view_pcr_dist.py
