+++ lib/view_pcr_dist_.py
